+++ packages/dash-table/src/dash-table/derived/cell/wrapperStyles.ts
@@ -2,7 +2,7 @@ import * as R from 'ramda';
 import { CSSProperties } from 'react';
 
 import { memoizeOneFactory } from 'core/memoizer';
-import { Data, VisibleColumns, IViewportOffset } from 'dash-table/components/Table/props';
+import { Data, VisibleColumns, IViewportOffset, SelectedCells, ICellCoordinates } from 'dash-table/components/Table/props';
 import { IConvertedStyle } from '../style';
 import { BORDER_PROPERTIES_AND_FRAGMENTS } from '../edges/type';
 
@@ -12,9 +12,10 @@ function getter(
     columns: VisibleColumns,
     columnStyles: IConvertedStyle[],
     data: Data,
-    offset: IViewportOffset
+    offset: IViewportOffset,
+    selectedCells: SelectedCells
 ): Style[][] {
-    return R.addIndex<any, Style[]>(R.map)((datum, index) => R.map(column => {
+    return R.addIndex<any, Style[]>(R.map)((datum, index) => R.addIndex<any, Style>(R.map)((column, columnIndex) => {
         const relevantStyles = R.map(
             s => s.style,
             R.filter<IConvertedStyle>(
@@ -25,7 +26,11 @@ function getter(
                 columnStyles
             )
         );
-
+        const matchCell = (cell: ICellCoordinates) => cell.row === index && cell.column === columnIndex;
+        const isSelectedCell: boolean = R.any(matchCell)(selectedCells);
+        if (isSelectedCell) {
+            relevantStyles.push({backgroundColor:  'var(--selected-background)'});
+        }
         return relevantStyles.length ?
             R.omit(
                 BORDER_PROPERTIES_AND_FRAGMENTS,