+++ packages/dash-table/CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [#434](https://github.com/plotly/dash-table/issues/434)
 - Fix CSS borders propeties overwrite style_* borders properties. 
 
+[#435](https://github.com/plotly/dash-table/issues/434)
+- selected_cells background color is set through styling pipeline / derivations. 
+
 ## [3.7.0] - 2019-05-15
 ### Added
 [#397](https://github.com/plotly/dash-table/pull/397), [#410](https://github.com/plotly/dash-table/pull/410)